+++ src/entity/product/itemForm/index.ts
@@ -1,4 +1,4 @@
-export { default as ItemFormHeader } from './ui/ItemFormHeader';
+export { default as ItemFormProgressBar } from './ui/ItemFormProgressBar';
 export { default as ImageUploader } from './ui/ImageUploader';
 export { default as NextButton } from './ui/NextButton';
 export { default as LastStepButton } from './ui/LastStepButton';
