+++ src/shared/ui/index.ts
@@ -3,3 +3,4 @@ export { Input } from './Input';
 export { Dropdown } from './Dropdown';
 export { ErrorMessage } from './ErrorMessage';
 export { Footer } from './Footer';
+export { Header } from './Header';
